+++ BabaSource/Editor/Utils/GridHelpers.cs
@@ -15,16 +15,16 @@ namespace Editor.Utils
         /// <param name="min"></param>
         /// <param name="max"></param>
         /// <returns></returns>
-        public static List<string> GetColumnHeaders(uint min, uint max, uint forceLength = 0)
+        public static List<string> GetColumnHeaders(int min, int max, int forceLength = 0)
         {
             var magnitude = (int)Math.Log(max, 26) + 1;
             var lines = " ".Repeat(max - min)
-                .Select((c, i) => EnumerableExtensions.ToColString((uint)i + min).PadLeft(magnitude))
+                .Select((c, i) => EnumerableExtensions.ToColString(i + min).PadLeft(magnitude))
                 .ZipMany().ToList();
 
             if (forceLength > 0)
             {
-                lines = new[] { "" }.Repeat((int)forceLength - lines.Count).Concat(lines).ToList();
+                lines = new[] { "" }.Repeat(forceLength - lines.Count).Concat(lines).ToList();
             }
 
             return lines;
@@ -37,7 +37,7 @@ namespace Editor.Utils
         /// <param name="min"></param>
         /// <param name="max"></param>
         /// <returns></returns>
-        public static List<string> GetRowHeaders(uint min, uint max, uint forceLength = 0)
+        public static List<string> GetRowHeaders(int min, int max, int forceLength = 0)
         {
             var magnitude = (int)Math.Log(max, 10) + 1;
             var lines = " ".Repeat(max - min)
